requirements.txt
python-dotenv
requests
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
flask-bootstrap
